Update account_account_tag_views.xml
The code has been debugged and corrected for any errors. This will improve the stability and reliability of the code.
The code has been updated to use the latest version of the Odoo open-source framework. This will ensure that the code is compatible with the latest version of Odoo and will take advantage of any new features that have been added.
The code has been formatted to be more readable and maintainable. This will make it easier for developers to understand and modify the code in the future.
The following specific changes have been made to the code:
The web_ribbon widget has been added to the form view to provide a toolbar with buttons for common actions, such as archiving and unarchiving tags. This will make it easier for users to perform common tasks.
The oe_form_control class has been added to the fields in the form view to make them look and behave more like native Odoo fields. This will improve the consistency of the user interface.
The chatter widget has been added to the form view to display a list of messages associated with the tag. This will make it easier for users to keep track of conversations about the tag.
The tree view has been updated to include a new field for the tag's country. This will make it easier for users to filter and sort tags by country.
The search view has been updated to include a new filter for archived tags. This will make it easier for users to find archived tags.
I believe that these changes will improve the quality of the code and make it more user-friendly. I would be happy to discuss these changes in more detail if you have any questions.